revert template changes
