--- templates/react-ecs/.gitattributes
@@ -1,3 +1,0 @@
-# suppress diffs for generated files
-**/pnpm-lock.yaml linguist-generated=true
-**/codegen/**/*.sol linguist-generated=true
