--- templates/react-ecs/.vscode/extensions.json
@@ -1,3 +1,0 @@
-{
-  "recommendations": ["NomicFoundation.hardhat-solidity"]
-}
