--- templates/react-ecs/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "mud-template-react-ecs",
-  "private": true,
-  "scripts": {
-    "build": "pnpm recursive run build",
-    "dev": "concurrently -n contracts,client -c cyan,magenta \"cd packages/contracts && pnpm run dev\" \"cd packages/client && pnpm run dev\"",
-    "dev:client": "pnpm --filter 'client' run dev",
-    "dev:contracts": "pnpm --filter 'contracts' dev",
-    "foundry:up": "curl -L https://foundry.paradigm.xyz | bash && bash $HOME/.foundry/bin/foundryup",
-    "mud:up": "pnpm mud set-version --tag main && pnpm install",
-    "prepare": "(forge --version || pnpm foundry:up)",
-    "test": "pnpm recursive run test"
-  },
-  "devDependencies": {
-    "@latticexyz/cli": "link:../../packages/cli",
-    "@typescript-eslint/eslint-plugin": "5.46.1",
-    "@typescript-eslint/parser": "5.46.1",
-    "concurrently": "^8.0.1",
-    "eslint": "8.29.0",
-    "rimraf": "^3.0.2",
-    "typescript": "5.1.6"
-  },
-  "engines": {
-    "node": "18.x",
-    "pnpm": "8.x"
-  }
-}
