--- templates/react-ecs/packages/client/.env
@@ -1,1 +1,0 @@
-VITE_CHAIN_ID=31337
