--- templates/react-ecs/packages/client/.eslintrc
@@ -1,7 +1,0 @@
-{
-  "extends": ["../../.eslintrc", "plugin:react/recommended", "plugin:react-hooks/recommended"],
-  "plugins": ["react", "react-hooks"],
-  "rules": {
-    "react/react-in-jsx-scope": "off"
-  }
-}
