--- templates/react-ecs/packages/client/.gitignore
@@ -1,3 +1,0 @@
-node_modules
-dist
-.DS_Store
