--- templates/react-ecs/packages/client/index.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>a minimal MUD client</title>
-  </head>
-  <body>
-    <div id="react-root"></div>
-    <script type="module" src="/src/index.tsx"></script>
-  </body>
-</html>
