--- templates/react-ecs/packages/client/package.json
@@ -1,38 +1,0 @@
-{
-  "name": "client",
-  "version": "0.0.0",
-  "private": true,
-  "license": "MIT",
-  "type": "module",
-  "scripts": {
-    "build": "vite build",
-    "dev": "wait-port localhost:8545 && vite",
-    "preview": "vite preview",
-    "test": "tsc --noEmit"
-  },
-  "dependencies": {
-    "@latticexyz/common": "link:../../../../packages/common",
-    "@latticexyz/dev-tools": "link:../../../../packages/dev-tools",
-    "@latticexyz/react": "link:../../../../packages/react",
-    "@latticexyz/recs": "link:../../../../packages/recs",
-    "@latticexyz/schema-type": "link:../../../../packages/schema-type",
-    "@latticexyz/services": "link:../../../../packages/services",
-    "@latticexyz/store-sync": "link:../../../../packages/store-sync",
-    "@latticexyz/utils": "link:../../../../packages/utils",
-    "@latticexyz/world": "link:../../../../packages/world",
-    "contracts": "workspace:*",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "rxjs": "7.5.5",
-    "viem": "1.14.0"
-  },
-  "devDependencies": {
-    "@types/react": "18.2.22",
-    "@types/react-dom": "18.2.7",
-    "@vitejs/plugin-react": "^3.1.0",
-    "eslint-plugin-react": "7.31.11",
-    "eslint-plugin-react-hooks": "4.6.0",
-    "vite": "^4.2.1",
-    "wait-port": "^1.0.4"
-  }
-}
