--- templates/react-ecs/packages/client/src/mud/supportedChains.ts
@@ -1,20 +1,0 @@
-/*
- * The supported chains.
- * By default, there are only two chains here:
- *
- * - mudFoundry, the chain running on anvil that pnpm dev
- *   starts by default. It is similar to the viem anvil chain
- *   (see https://viem.sh/docs/clients/test.html), but with the
- *   basefee set to zero to avoid transaction fees.
- * - latticeTestnet, our public test network.
- *
-
- */
-
-import { MUDChain, latticeTestnet, mudFoundry } from "@latticexyz/common/chains";
-
-/*
- * See https://mud.dev/tutorials/minimal/deploy#run-the-user-interface
- * for instructions on how to add networks.
- */
-export const supportedChains: MUDChain[] = [mudFoundry, latticeTestnet];
