--- templates/react-ecs/packages/client/tsconfig.json
@@ -1,19 +1,0 @@
-{
-  "compilerOptions": {
-    "types": ["vite/client"],
-    "target": "ESNext",
-    "useDefineForClassFields": true,
-    "module": "ESNext",
-    "lib": ["ESNext", "DOM"],
-    "moduleResolution": "Node",
-    "strict": true,
-    "resolveJsonModule": true,
-    "isolatedModules": true,
-    "esModuleInterop": true,
-    "noEmit": true,
-    "skipLibCheck": true,
-    "jsx": "react-jsx",
-    "jsxImportSource": "react"
-  },
-  "include": ["src"]
-}
