--- templates/react-ecs/packages/client/vite.config.ts
@@ -1,17 +1,0 @@
-import { defineConfig } from "vite";
-import react from "@vitejs/plugin-react";
-
-export default defineConfig({
-  plugins: [react()],
-  server: {
-    port: 3000,
-    fs: {
-      strict: false,
-    },
-  },
-  build: {
-    target: "es2022",
-    minify: true,
-    sourcemap: true,
-  },
-});
