--- templates/react-ecs/packages/contracts/.env
@@ -1,8 +1,0 @@
-# This .env file is for demonstration purposes only.
-#
-# This should usually be excluded via .gitignore and the env vars attached to
-# your deployment enviroment, but we're including this here for ease of local
-# development. Please do not commit changes to this file!
-#
-# Anvil default private key:
-PRIVATE_KEY=0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80
