--- templates/react-ecs/packages/contracts/.gitignore
@@ -1,9 +1,0 @@
-out/
-cache/
-node_modules/
-bindings/
-artifacts/
-broadcast/
-
-# Ignore MUD deploy artifacts
-deploys/**/*.json
