--- templates/react-ecs/packages/contracts/.prettierrc
@@ -1,8 +1,0 @@
-{
-  "plugins": ["prettier-plugin-solidity"],
-  "printWidth": 120,
-  "semi": true,
-  "tabWidth": 2,
-  "useTabs": false,
-  "bracketSpacing": true
-}
