--- templates/react-ecs/packages/contracts/.solhint.json
@@ -1,12 +1,0 @@
-{
-  "extends": ["solhint:recommended", "mud"],
-  "plugins": ["mud"],
-  "rules": {
-    "compiler-version": ["error", ">=0.8.0"],
-    "avoid-low-level-calls": "off",
-    "no-inline-assembly": "off",
-    "func-visibility": ["warn", { "ignoreConstructors": true }],
-    "no-empty-blocks": "off",
-    "no-complex-fallback": "off"
-  }
-}
