--- templates/react-ecs/packages/contracts/foundry.toml
@@ -1,26 +1,0 @@
-[profile.default]
-solc = "0.8.21"
-ffi = false
-fuzz_runs = 256
-optimizer = true
-optimizer_runs = 3000
-verbosity = 2
-src = "src"
-test = "test"
-out = "out"
-allow_paths = [
-  # pnpm symlinks to the project root's node_modules
-  "../../node_modules",
-  # template uses linked mud packages from within the mud monorepo
-  "../../../../packages",
-  # projects created from this template and using linked mud packages
-  "../../../mud/packages",
-]
-extra_output_files = [
-  "abi",
-  "evm.bytecode"
-]
-fs_permissions = [{ access = "read", path = "./"}]
-
-[profile.lattice-testnet]
-eth_rpc_url = "https://follower.testnet-chain.linfra.xyz"
