--- templates/react-ecs/packages/contracts/mud.config.ts
@@ -1,10 +1,0 @@
-import { mudConfig } from "@latticexyz/world/register";
-
-export default mudConfig({
-  tables: {
-    Counter: {
-      keySchema: {},
-      valueSchema: "uint32",
-    },
-  },
-});
