--- templates/react-ecs/packages/contracts/package.json
@@ -1,36 +1,0 @@
-{
-  "name": "contracts",
-  "version": "0.0.0",
-  "private": true,
-  "license": "MIT",
-  "scripts": {
-    "build": "pnpm run build:mud && pnpm run build:abi && pnpm run build:abi-ts",
-    "build:abi": "forge build",
-    "build:abi-ts": "mud abi-ts",
-    "build:mud": "rimraf src/codegen && mud tablegen && mud worldgen",
-    "deploy:local": "pnpm run build && mud deploy",
-    "deploy:testnet": "pnpm run build && mud deploy --profile=lattice-testnet",
-    "dev": "pnpm mud dev-contracts",
-    "lint": "pnpm run prettier && pnpm run solhint",
-    "prettier": "prettier --write 'src/**/*.sol'",
-    "solhint": "solhint --config ./.solhint.json 'src/**/*.sol' --fix",
-    "test": "tsc --noEmit && mud test"
-  },
-  "dependencies": {
-    "@latticexyz/cli": "link:../../../../packages/cli",
-    "@latticexyz/schema-type": "link:../../../../packages/schema-type",
-    "@latticexyz/store": "link:../../../../packages/store",
-    "@latticexyz/world": "link:../../../../packages/world",
-    "@latticexyz/world-modules": "link:../../../../packages/world-modules"
-  },
-  "devDependencies": {
-    "@types/node": "^18.15.11",
-    "ds-test": "https://github.com/dapphub/ds-test.git#e282159d5170298eb2455a6c05280ab5a73a4ef0",
-    "forge-std": "https://github.com/foundry-rs/forge-std.git#74cfb77e308dd188d2f58864aaf44963ae6b88b1",
-    "prettier": "^2.6.2",
-    "prettier-plugin-solidity": "^1.0.0-beta.19",
-    "solhint": "^3.3.7",
-    "solhint-config-mud": "file:../../../../packages/solhint-config-mud",
-    "solhint-plugin-mud": "file:../../../../packages/solhint-plugin-mud"
-  }
-}
