--- templates/react-ecs/packages/contracts/remappings.txt
@@ -1,3 +1,0 @@
-ds-test/=node_modules/ds-test/src/
-forge-std/=node_modules/forge-std/src/
-@latticexyz/=node_modules/@latticexyz/
