--- templates/react-ecs/packages/contracts/src/codegen/index.sol
@@ -1,6 +1,0 @@
-// SPDX-License-Identifier: MIT
-pragma solidity >=0.8.21;
-
-/* Autogenerated file. Do not edit manually. */
-
-import { Counter, CounterTableId } from "./tables/Counter.sol";
