--- templates/react-ecs/packages/contracts/src/codegen/world/IIncrementSystem.sol
@@ -1,12 +1,0 @@
-// SPDX-License-Identifier: MIT
-pragma solidity >=0.8.21;
-
-/* Autogenerated file. Do not edit manually. */
-
-/**
- * @title IIncrementSystem
- * @dev This interface is automatically generated from the corresponding system contract. Do not edit manually.
- */
-interface IIncrementSystem {
-  function increment() external returns (uint32);
-}
