--- templates/react-ecs/packages/contracts/tsconfig.json
@@ -1,13 +1,0 @@
-// Visit https://aka.ms/tsconfig.json for all config options
-{
-  "compilerOptions": {
-    "target": "ES2020",
-    "module": "commonjs",
-    "strict": true,
-    "resolveJsonModule": true,
-    "esModuleInterop": true,
-    "skipLibCheck": true,
-    "forceConsistentCasingInFileNames": true,
-    "moduleResolution": "node"
-  }
-}
