--- templates/react-ecs/packages/contracts/worlds.json
@@ -1,5 +1,0 @@
-{
-  "31337": {
-    "address": "0x5FbDB2315678afecb367f032d93F642f64180aa3"
-  }
-}
