--- templates/react-ecs/packages/contracts/worlds.json.d.ts
@@ -1,2 +1,0 @@
-declare const worlds: Partial<Record<string, { address: string; blockNumber?: number }>>;
-export default worlds;
