+++ templates/react/packages/client/package.json
@@ -14,6 +14,7 @@
     "@latticexyz/common": "link:../../../../packages/common",
     "@latticexyz/dev-tools": "link:../../../../packages/dev-tools",
     "@latticexyz/react": "link:../../../../packages/react",
+    "@latticexyz/recs": "link:../../../../packages/recs",
     "@latticexyz/schema-type": "link:../../../../packages/schema-type",
     "@latticexyz/services": "link:../../../../packages/services",
     "@latticexyz/store-sync": "link:../../../../packages/store-sync",
