+++ templates/react/packages/client/src/index.tsx
@@ -28,6 +28,7 @@ setup().then(async (result) => {
       worldAddress: result.network.worldContract.address,
       worldAbi: result.network.worldContract.abi,
       write$: result.network.write$,
+      recsWorld: result.network.world,
     });
   }
 });
