+++ templates/react/packages/client/src/mud/createClientComponents.ts
