+++ templates/react/packages/client/src/mud/setup.ts
@@ -2,6 +2,7 @@
  * This file sets up all the definitions required for a MUD client.
  */
 
+import { createClientComponents } from "./createClientComponents";
 import { createSystemCalls } from "./createSystemCalls";
 import { setupNetwork } from "./setupNetwork";
 
@@ -9,10 +10,12 @@ export type SetupResult = Awaited<ReturnType<typeof setup>>;
 
 export async function setup() {
   const network = await setupNetwork();
-  const systemCalls = createSystemCalls(network);
+  const components = createClientComponents(network);
+  const systemCalls = createSystemCalls(network, components);
 
   return {
     network,
+    components,
     systemCalls,
   };
 }
