+++ templates/react/packages/client/src/mud/world.ts
