+++ templates/react/packages/contracts/mud.config.ts
@@ -2,12 +2,9 @@ import { mudConfig } from "@latticexyz/world/register";
 
 export default mudConfig({
   tables: {
-    Tasks: {
-      valueSchema: {
-        createdAt: "uint256",
-        completedAt: "uint256",
-        description: "string",
-      },
+    Counter: {
+      keySchema: {},
+      valueSchema: "uint32",
     },
   },
 });
