+++ templates/react/packages/contracts/script/PostDeploy.s.sol
@@ -6,7 +6,6 @@ import { console } from "forge-std/console.sol";
 import { StoreSwitch } from "@latticexyz/store/src/StoreSwitch.sol";
 
 import { IWorld } from "../src/codegen/world/IWorld.sol";
-import { Tasks, TasksData } from "../src/codegen/index.sol";
 
 contract PostDeploy is Script {
   function run(address worldAddress) external {
@@ -19,14 +18,11 @@ contract PostDeploy is Script {
     // Start broadcasting transactions from the deployer account
     vm.startBroadcast(deployerPrivateKey);
 
-    // We can set table records directly
-    Tasks.set("1", TasksData({ description: "Walk the dog", createdAt: block.timestamp, completedAt: 0 }));
+    // ------------------ EXAMPLES ------------------
 
-    // Or we can call our own systems
-    IWorld(worldAddress).addTask("Take out the trash");
-
-    bytes32 key = IWorld(worldAddress).addTask("Do the dishes");
-    IWorld(worldAddress).completeTask(key);
+    // Call increment on the world via the registered function selector
+    uint32 newValue = IWorld(worldAddress).increment();
+    console.log("Increment via IWorld:", newValue);
 
     vm.stopBroadcast();
   }
