+++ templates/react/packages/contracts/src/codegen/index.sol
@@ -3,4 +3,4 @@ pragma solidity >=0.8.21;
 
 /* Autogenerated file. Do not edit manually. */
 
-import { Tasks, TasksData, TasksTableId } from "./tables/Tasks.sol";
+import { Counter, CounterTableId } from "./tables/Counter.sol";
