+++ templates/react/packages/contracts/src/codegen/tables/Counter.sol
