--- templates/react/packages/contracts/src/codegen/world/ITaskSystem.sol
@@ -1,12 +1,0 @@
-// SPDX-License-Identifier: MIT
-pragma solidity >=0.8.21;
-
-/* Autogenerated file. Do not edit manually. */
-
-/**
- * @title ITaskSystem
- * @dev This interface is automatically generated from the corresponding system contract. Do not edit manually.
- */
-interface ITaskSystem {
-  function addTask(string memory label) external returns (bytes32 key);
-}
