+++ templates/react/packages/contracts/src/codegen/world/IWorld.sol
@@ -5,7 +5,7 @@ pragma solidity >=0.8.21;
 
 import { IBaseWorld } from "@latticexyz/world/src/codegen/interfaces/IBaseWorld.sol";
 
-import { ITasksSystem } from "./ITasksSystem.sol";
+import { IIncrementSystem } from "./IIncrementSystem.sol";
 
 /**
  * @title IWorld
@@ -13,6 +13,6 @@ import { ITasksSystem } from "./ITasksSystem.sol";
  * that are dynamically registered in the World during deployment.
  * @dev This is an autogenerated file; do not edit manually.
  */
-interface IWorld is IBaseWorld, ITasksSystem {
+interface IWorld is IBaseWorld, IIncrementSystem {
 
 }
