+++ templates/react/packages/contracts/worlds.json
@@ -1,5 +1,5 @@
 {
   "31337": {
-    "address": "0x6d584a9c0bd815104ab1fe92087c74450f7845d4"
+    "address": "0x5FbDB2315678afecb367f032d93F642f64180aa3"
   }
 }
